data-science/python/be-analytic-steamDB/requirements.txt
beautifulsoup4
zenrows
google-cloud-bigquery
